Refactor into a separate class to avoid polluting the Util namespace
