+++ XIV on Mac/AppDelegate.swift
@@ -26,7 +26,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             macButton.state = .off
             winButton.state = .on
         }
-        let discord = Util.Discord()
+        let discord = DiscordRichPresence()
         if (discord.connected) {
             discord.setPresence()
         }
