+++ XIV on Mac/Util.swift
@@ -7,8 +7,6 @@
 
 import Cocoa
 
-import SwordRPC
-
 struct Util {
     @available(*, unavailable) private init() {}
     
@@ -93,32 +91,6 @@ struct Util {
     static func killWine(logger: NSTextView?) {
         launch(exec: wineserver, args: ["-k"])
     }
-    
-    class Discord {
-        var discordAppId = "925647215421173820"
-        var version = Bundle.main.infoDictionary!["CFBundleShortVersionString"] as? String ?? "1.0.0"
-        var state = "https://www.xivmac.com/"
-        var iconName = "appicon"
-        
-        var rpc:SwordRPC
-        var richPresence:RichPresence
-        var connected:Bool
-        var details:String
-        init() {
-            self.rpc = SwordRPC.init(appId: discordAppId)
-            self.connected = self.rpc.connect()
-            self.richPresence = RichPresence()
-            self.details = "Version " + version
-        }
-        
-        func setPresence() {
-            self.richPresence.assets.largeImage = iconName
-            self.richPresence.details = details
-            self.richPresence.state = state
-            rpc.setPresence(self.richPresence)
-        }
-    }
-    
     
     class DXVK: Codable {
         static let settingKey = "DXVK_OPTIONS"
